agent.py split into mixins
